version-1.1

+++ src/com/trainingshare/servlet/NewMyTitleServlet.java
@@ -50,7 +50,7 @@ public class NewMyTitleServlet extends HttpServlet {
 		String location = request.getParameter("roomnumber");
 		
 		//如果删除标志不为空，则执行删除主题操作
-		if(null!=removeFlag || removeFlag.equals("true"))
+		if(null!=removeFlag && removeFlag.equals("true"))
 		{
 			if(new DBConnect().DeleteMemberPerson(membersId, title))
 				response.getWriter().write("remove_success");
@@ -59,51 +59,46 @@ public class NewMyTitleServlet extends HttpServlet {
 		}
 		else//负责执行新建或更新操作
 		{
-			if(null!=title && !title.equals("") && null!=titleoriginal
-					   && null != time && !time.equals("")
-					   && null != location && !location.equals(""))
+			ActivityContentBean acb = new ActivityContentBean();
+			acb.setMembersId(membersId);
+			acb.setMemberId(dbc.GetMemberIdByName(new String[]{username}).get(0));
+			acb.setTitle(title);
+			acb.setMeetingRoomId(dbc.GetMeetingRoomIdByName(location));
+			acb.setStartTime(time.replace("/", "-"));
+			acb.setFilePath("");
+			acb.setUploadFlag("0");
+			acb.setRemark("");
+			//如果原标题为空，则说明此时正在新建标题业务，应该执行插入操作
+			if(titleoriginal.equals(""))
+			{
+				//判断该标题是否已存在
+				if(!dbc.CheckActivityContentExistorNot(membersId,title))
+				{
+					if(dbc.AddNewAcivityContent(acb) == true)
 					{
-						ActivityContentBean acb = new ActivityContentBean();
-						acb.setMembersId(membersId);
-						acb.setMemberId(dbc.GetMemberIdByName(new String[]{username}).get(0));
-						acb.setTitle(title);
-						acb.setMeetingRoomId(dbc.GetMeetingRoomIdByName(location));
-						acb.setStartTime(time.replace("/", "-"));
-						acb.setFilePath("");
-						acb.setUploadFlag("0");
-						acb.setRemark("");
-						//如果原标题为空，则说明此时正在新建标题业务，应该执行插入操作
-						if(titleoriginal.equals(""))
-						{
-							if(dbc.AddNewAcivityContent(acb) == true)
-							{
-								//request.setAttribute("membersId", Integer.toString(membersId));
-								//request.setAttribute("activityContentList", dbc.GetActivityContentByMembersId(membersId));
-								//request.getRequestDispatcher("ActivityContent.jsp").forward(request,response);
-								response.getWriter().write("save_success");
-							}
-							else
-							{
-								response.getWriter().write("save_failed");
-							}
-						}
-						else//否则执行更新操作
-						{
-							if(dbc.UpdateActivityContentByTitle(titleoriginal, acb))
-							{
-								response.getWriter().write("edit_success");
-							}
-							else
-							{
-								response.getWriter().write("edit_failed");
-							}
-						}
+						response.getWriter().write("save_success");
 					}
 					else
 					{
-						response.getWriter().write("failed");
+						response.getWriter().write("save_failed");
 					}
+				}
+				else
+				{
+					response.getWriter().write("title exist");
+				}
+			}
+			else//否则执行更新操作
+			{
+				if(dbc.UpdateActivityContentByTitle(titleoriginal, acb))
+				{
+					response.getWriter().write("edit_success");
+				}
+				else
+				{
+					response.getWriter().write("edit_failed");
+				}
+			}
 		}
 	}
-
 }
